pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[assemblyai,openai,rime,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-silero~=0.2",
    "python-dotenv>=1.1.1",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "livekit>=0.15.0",
    "aiohttp>=3.8.0",
]
